requirements.txt
discord.py==2.3.2
datasets
sentence-transformers
faiss-cpu
requests
python-dotenv
numpy
torch
transformers
beautifulsoup4
aiohttp
wikipedia-api
lxml
psutil